==> notebooks/8.Plotting.ipynb <==
### Import Packages

import pandas as pd
from modules.utils import load_json
import warnings
warnings.filterwarnings("ignore")

### Prepare Data For PCA

df = pd.read_csv('../BENCHMARKING/Selected_Biomarker_Panels.csv') 
df.rename(columns = {
        'randomforest_feature_importance': 'RF-FI',
        'xgb_feature_importance': 'XGB-FI',
        'rf_permutation_feature_importance': 'RF-PFI',
        'xgb_permutation_feature_importance': 'XGB-PFI'
    }, inplace = True) 
df = df[df.OmicsLevel == 'TripleOmics']
df.head()

cohorts = ['ROSMAP', 'BRCA', 'MayoRNASeq'] 
rankers = df.columns[4:]
Selection_Sets = {cohort: {ranker: [] for ranker in rankers} for cohort in cohorts}
for cohort in cohorts:
    df_cohort = df[df.Cohort==cohort]
    all_features = df_cohort.Feature.to_list()
    
    for k in range(10,101,10):
        for ranker in rankers:
            feature_set = df_cohort[df_cohort[ranker] <= k].Feature.to_list() 
            for feat in all_features: 
                if feat in feature_set:
                    Selection_Sets[cohort][ranker].append(1)
                else:
                    Selection_Sets[cohort][ranker].append(0)

# Save PCA data
for cohort in cohorts:
    df2 = pd.DataFrame(Selection_Sets[cohort])
    df2 = df2[df2.sum(axis=1)>0]
    df2.to_csv(f'../BENCHMARKING/{cohort}_PCA_data.csv', index = False)

### Prepare Features For Heatmap

# select top features for each Cohort: Features that got ranked amongst top 20 in atleast 1 ranker
top_features = {"ROSMAP": [], "BRCA":[], "MayoRNASeq": []}
for cohort in top_features:
    cohort_df = df[df["Cohort"]==cohort]
    for feat in cohort_df.Feature.to_list():
        if sum(cohort_df[cohort_df.Feature==feat].iloc[0,4:] <= 20) > 5:
            top_features[cohort].append(feat) 

for key, value in top_features.items():
    print(key, len(value))

def find_name(x):
    result = None
    for y in names:
        if x in feature_names[names[y]]:  
            result = y
            break
    return result


# prepare Cohort data
cohort_data_filtered = {}
for cohort in top_features:
    cohort_df = df[df["Cohort"]==cohort][df.Feature.isin(top_features[cohort])].drop(columns = df.columns[1:4])
    cohort_df["rank"] = cohort_df.iloc[:,1:].sum(axis=1)
    cohort_df.sort_values(by="rank", ascending = True)
    cohort_df.drop(columns = "rank", inplace = True) 

    feature_names = load_json(f"../results/{cohort}/featurenames.json") 

    if cohort == "ROSMAP" or cohort=="BRCA":
        names = {"miRNA": "1", "mRNA":"2", "Meth":"3"}
    else:
        names = {"Metab": "1", "Prot":"2", "mRNA":"3"}
    cohort_df.insert(0,"FeatureType", cohort_df['Feature'].apply(find_name))
    cohort_df.to_csv(f"../BENCHMARKING/{cohort}_Ranks_data.csv", index = False)

### Prepare Crosvalidation Results for Overleap

df = pd.read_csv('../BENCHMARKING/Cross_Validation_Results.csv') 
df = df[df.featureSelector=='NONE'] 
df.drop(columns = ['numFeatures','featureSelector','MeanLR_PLUS', 'StdLR_PLUS', 'MeanLR_MINUS', 'StdLR_MINUS'], inplace = True)

for metric in ['Accuracy', 'Precision', 'Recall', 'F1', 'AUC', 'Specificity', 'NPV']:
    df[metric] = df.apply(lambda x: f'{x[f"Mean{metric}"]:.2f} ± {x[f"Std{metric}"]:.2f}' , axis = 1)


df.drop(columns = [column for column in df.columns if ('Mean' in column or 'Std' in column)], inplace =True)

df[df.OmicsLevel=='TripleOmics'].to_latex()
